Move setting pfp function to an object
To reduce duplicate code, I moved the function that set an a profile picture (an ImageView) based on users/uid/profile_pic to a separate class. Also, I renamed the function to setImageViewToProfilePic from setProfilePicture.
