+++ Project/app/src/main/java/com/example/cmpt362project/MainActivity.kt
@@ -1,10 +1,7 @@
 package com.example.cmpt362project
 
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
 import android.os.Bundle
 import android.view.Menu
-import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
@@ -14,15 +11,13 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
-import com.google.android.material.floatingactionbutton.FloatingActionButton
+import com.example.cmpt362project.utility.ImageUtility
 import com.google.android.material.navigation.NavigationView
-import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
-import com.google.firebase.storage.ktx.storage
 
 class MainActivity : AppCompatActivity() {
 
@@ -56,7 +51,7 @@ class MainActivity : AppCompatActivity() {
                     val userData = it.value as Map<*, *>
                     navView.getHeaderView(0).findViewById<TextView>(R.id.drawer_header_username).text = userData["username"].toString()
                     navView.getHeaderView(0).findViewById<TextView>(R.id.drawer_header_email).text = userData["email"].toString()
-                    setProfilePicture(navView.getHeaderView(0).findViewById(R.id.drawer_header_profile_pic))
+                    ImageUtility.setImageViewToProfilePic(navView.getHeaderView(0).findViewById(R.id.drawer_header_profile_pic))
                 }
 
             }
@@ -72,29 +67,5 @@ class MainActivity : AppCompatActivity() {
     override fun onSupportNavigateUp(): Boolean {
         val navController = findNavController(R.id.nav_host_fragment)
         return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
-    }
-
-    private fun setProfilePicture(imageView: ImageView) {
-        val storage = Firebase.storage.reference
-        val auth = Firebase.auth
-        val user = auth.currentUser
-
-        database.child("users").child(user!!.uid).child("profile_pic").get()
-            .addOnSuccessListener { pathToProfilePic ->
-
-                val pathReference = storage.child(pathToProfilePic.value as String)
-                val oneMegabyte: Long = 1024 * 1024
-
-                pathReference.getBytes(oneMegabyte)
-                    .addOnSuccessListener(this) {
-                        val bitmap: Bitmap? = BitmapFactory.decodeByteArray(it, 0, it.size)
-                        if (bitmap != null) imageView.setImageBitmap(bitmap)
-                    } .addOnFailureListener(this) {
-                        println("MainActivity setProfilePicture: Failed to download profile pic")
-                    }
-            }
-            .addOnFailureListener {
-                println("MainActivity setProfilePicture: Failed to get value users/uid/profile_pic")
-            }
     }
 }
