+++ Project/app/src/main/java/com/example/cmpt362project/ui/settings/UserProfileActivity.kt
@@ -5,36 +5,27 @@ import android.content.DialogInterface
 import android.content.Intent
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import android.net.Uri
 import android.os.Bundle
 import android.provider.MediaStore
 import android.view.View
 import android.widget.EditText
 import android.widget.ImageView
-import android.widget.TextView
 import android.widget.Toast
-import androidx.activity.result.ActivityResult
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.example.cmpt362project.R
-import com.google.android.material.textfield.TextInputEditText
+import com.example.cmpt362project.utility.ImageUtility
 import com.google.android.material.textfield.TextInputLayout
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
-import com.google.firebase.storage.FirebaseStorage
-import com.google.firebase.storage.StorageException.ERROR_OBJECT_NOT_FOUND
-import com.google.firebase.storage.StorageReference
 import com.google.firebase.storage.ktx.storage
 import java.io.ByteArrayOutputStream
-import java.io.File
-import java.io.FileOutputStream
 
 class UserProfileActivity : AppCompatActivity() {
 
@@ -106,7 +97,7 @@ class UserProfileActivity : AppCompatActivity() {
 //            pictureView.setImageBitmap(placeholderImage)
 //        }
 
-        setProfilePicture(pictureView)
+        ImageUtility.setImageViewToProfilePic(pictureView)
 
 
         // Initialize the gallery activity
@@ -183,29 +174,5 @@ class UserProfileActivity : AppCompatActivity() {
                 .addOnFailureListener(this) { println("Failure upload") }
                 .addOnCompleteListener(this) { println("Upload complete!") }
         }
-    }
-
-    private fun setProfilePicture(imageView: ImageView) {
-        val storage = Firebase.storage.reference
-        val auth = Firebase.auth
-        val user = auth.currentUser
-
-        database.child("users").child(user!!.uid).child("profile_pic").get()
-            .addOnSuccessListener { pathToProfilePic ->
-
-                val pathReference = storage.child(pathToProfilePic.value as String)
-                val oneMegabyte: Long = 1024 * 1024
-
-                pathReference.getBytes(oneMegabyte)
-                    .addOnSuccessListener(this) {
-                        val bitmap: Bitmap? = BitmapFactory.decodeByteArray(it, 0, it.size)
-                        if (bitmap != null) imageView.setImageBitmap(bitmap)
-                    } .addOnFailureListener(this) {
-                        println("MainActivity setProfilePicture: Failed to download profile pic")
-                    }
-            }
-            .addOnFailureListener {
-                println("MainActivity setProfilePicture: Failed to get value users/uid/profile_pic")
-            }
     }
 }
